+++ src/assets/js/store/modules/workouts/actions.js
@@ -17,5 +17,9 @@ export default {
 
   addExercise(context, payload) {
     context.commit("addExercise", payload);
+  },
+
+  removeExercise(context, payload) {
+    context.commit("removeExercise", payload);
   }
 };